# jsonpatchx/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "AddOp": {
        "additionalProperties": false,
        "description": "RFC 6902 add operation.",
        "properties": {
          "op": {
            "const": "add",
            "default": "add",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JSONValue-Input"
          }
        },
        "required": [
          "op",
          "path",
          "value"
        ],
        "title": "Add operation",
        "type": "object"
      },
      "Config_Patch_Document": {
        "description": "Array of patch operations for Config.",
        "items": {
          "$ref": "#/components/schemas/Config_Patch_Operation"
        },
        "title": "Config Patch Document",
        "type": "array"
      },
      "Config_Patch_Operation": {
        "description": "Discriminated union of patch operations for Config.",
        "discriminator": {
          "mapping": {
            "add": "#/components/schemas/AddOp",
            "copy": "#/components/schemas/CopyOp",
            "move": "#/components/schemas/MoveOp",
            "remove": "#/components/schemas/RemoveOp",
            "replace": "#/components/schemas/ReplaceOp",
            "test": "#/components/schemas/TestOp",
            "toggle": "#/components/schemas/ToggleOp"
          },
          "propertyName": "op"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/AddOp"
          },
          {
            "$ref": "#/components/schemas/CopyOp"
          },
          {
            "$ref": "#/components/schemas/MoveOp"
          },
          {
            "$ref": "#/components/schemas/RemoveOp"
          },
          {
            "$ref": "#/components/schemas/ReplaceOp"
          },
          {
            "$ref": "#/components/schemas/TestOp"
          },
          {
            "$ref": "#/components/schemas/ToggleOp"
          }
        ],
        "title": "Config Patch Operation"
      },
      "CopyOp": {
        "additionalProperties": false,
        "description": "RFC 6902 copy operation.",
        "properties": {
          "from": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "From",
            "type": "string"
          },
          "op": {
            "const": "copy",
            "default": "copy",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "from",
          "op",
          "path"
        ],
        "title": "Copy operation",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JSONArray_JSONValue_-Input": {
        "items": {
          "$ref": "#/components/schemas/JSONValue-Input"
        },
        "title": "JSON array",
        "type": "array"
      },
      "JSONArray_JSONValue_-Output": {
        "items": {
          "$ref": "#/components/schemas/JSONValue-Output"
        },
        "title": "JSON array",
        "type": "array"
      },
      "JSONBoolean": {
        "title": "JSON boolean",
        "type": "boolean"
      },
      "JSONNull": {
        "title": "JSON null",
        "type": "null"
      },
      "JSONNumber": {
        "anyOf": [
          {
            "type": "integer"
          },
          {
            "type": "number"
          }
        ],
        "description": "integer or finite float (no NaN/Infinity).",
        "title": "JSON number"
      },
      "JSONObject_JSONValue_-Input": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JSONValue-Input"
        },
        "title": "JSON object",
        "type": "object"
      },
      "JSONObject_JSONValue_-Output": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JSONValue-Output"
        },
        "title": "JSON object",
        "type": "object"
      },
      "JSONString": {
        "title": "JSON string",
        "type": "string"
      },
      "JSONValue-Input": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/JSONBoolean"
          },
          {
            "$ref": "#/components/schemas/JSONNumber"
          },
          {
            "$ref": "#/components/schemas/JSONString"
          },
          {
            "$ref": "#/components/schemas/JSONNull"
          },
          {
            "$ref": "#/components/schemas/JSONArray_JSONValue_-Input"
          },
          {
            "$ref": "#/components/schemas/JSONObject_JSONValue_-Input"
          }
        ],
        "title": "JSON value"
      },
      "JSONValue-Output": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/JSONBoolean"
          },
          {
            "$ref": "#/components/schemas/JSONNumber"
          },
          {
            "$ref": "#/components/schemas/JSONString"
          },
          {
            "$ref": "#/components/schemas/JSONNull"
          },
          {
            "$ref": "#/components/schemas/JSONArray_JSONValue_-Output"
          },
          {
            "$ref": "#/components/schemas/JSONObject_JSONValue_-Output"
          }
        ],
        "title": "JSON value"
      },
      "MedicalRecord": {
        "additionalProperties": false,
        "properties": {
          "diagnosis": {
            "title": "Diagnosis",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "diagnosis"
        ],
        "title": "MedicalRecord",
        "type": "object"
      },
      "MedicalRecord_Patch_Document": {
        "description": "Array of patch operations for MedicalRecord. Applied to model_dump() and re-validated against the model schema.",
        "items": {
          "$ref": "#/components/schemas/MedicalRecord_Patch_Operation"
        },
        "title": "MedicalRecord Patch Document",
        "type": "array",
        "x-target-model": "MedicalRecord"
      },
      "MedicalRecord_Patch_Operation": {
        "description": "Discriminated union of patch operations for MedicalRecord.",
        "discriminator": {
          "mapping": {
            "add": "#/components/schemas/AddOp",
            "copy": "#/components/schemas/CopyOp",
            "move": "#/components/schemas/MoveOp",
            "remove": "#/components/schemas/RemoveOp",
            "replace": "#/components/schemas/ReplaceOp",
            "test": "#/components/schemas/TestOp",
            "toggle": "#/components/schemas/ToggleOp"
          },
          "propertyName": "op"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/AddOp"
          },
          {
            "$ref": "#/components/schemas/CopyOp"
          },
          {
            "$ref": "#/components/schemas/MoveOp"
          },
          {
            "$ref": "#/components/schemas/RemoveOp"
          },
          {
            "$ref": "#/components/schemas/ReplaceOp"
          },
          {
            "$ref": "#/components/schemas/TestOp"
          },
          {
            "$ref": "#/components/schemas/ToggleOp"
          }
        ],
        "title": "MedicalRecord Patch Operation"
      },
      "MoveOp": {
        "additionalProperties": false,
        "description": "RFC 6902 move operation.",
        "properties": {
          "from": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "From",
            "type": "string"
          },
          "op": {
            "const": "move",
            "default": "move",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "from",
          "op",
          "path"
        ],
        "title": "Move operation",
        "type": "object"
      },
      "RemoveOp": {
        "additionalProperties": false,
        "description": "RFC 6902 remove operation.",
        "properties": {
          "op": {
            "const": "remove",
            "default": "remove",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "op",
          "path"
        ],
        "title": "Remove operation",
        "type": "object"
      },
      "ReplaceOp": {
        "additionalProperties": false,
        "description": "RFC 6902 replace operation.",
        "properties": {
          "op": {
            "const": "replace",
            "default": "replace",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JSONValue-Input"
          }
        },
        "required": [
          "op",
          "path",
          "value"
        ],
        "title": "Replace operation",
        "type": "object"
      },
      "TestOp": {
        "additionalProperties": false,
        "description": "RFC 6902 test operation.",
        "properties": {
          "op": {
            "const": "test",
            "default": "test",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JSONValue-Input"
          }
        },
        "required": [
          "op",
          "path",
          "value"
        ],
        "title": "Test operation",
        "type": "object"
      },
      "ToggleOp": {
        "additionalProperties": false,
        "properties": {
          "op": {
            "const": "toggle",
            "default": "toggle",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "op",
          "path"
        ],
        "title": "ToggleOp",
        "type": "object"
      },
      "User": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "User",
        "type": "object"
      },
      "User_Patch_Document": {
        "description": "Array of patch operations for User. Applied to model_dump() and re-validated against the model schema.",
        "items": {
          "$ref": "#/components/schemas/User_Patch_Operation"
        },
        "title": "User Patch Document",
        "type": "array",
        "x-target-model": "User"
      },
      "User_Patch_Operation": {
        "description": "Discriminated union of patch operations for User.",
        "discriminator": {
          "mapping": {
            "toggle": "#/components/schemas/ToggleOp"
          },
          "propertyName": "op"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/ToggleOp"
          }
        ],
        "title": "User Patch Operation"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "jsonpatchx openapi snapshot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/configs/{config_id}": {
      "patch": {
        "operationId": "patch_config_configs__config_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "title": "Config Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config_Patch_Document"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONValue-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Config"
      }
    },
    "/configs/{config_id}/dep": {
      "patch": {
        "operationId": "patch_config_dep_configs__config_id__dep_patch",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "title": "Config Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DotConfigPatch_Patch_Document",
                "title": "Patch"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/DotConfigPatch_Patch_Document"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONValue-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Config Dep"
      }
    },
    "/records/{record_id}": {
      "patch": {
        "operationId": "patch_record_records__record_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecord_Patch_Document"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Record"
      }
    },
    "/records/{record_id}/dep": {
      "patch": {
        "operationId": "patch_record_dep_records__record_id__dep_patch",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecord_Patch_Document",
                "title": "Patch"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecord_Patch_Document"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Record Dep"
      }
    },
    "/users/{user_id}": {
      "patch": {
        "operationId": "patch_user_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User_Patch_Document"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch User"
      }
    }
  }
}